refactor(website): extract SitePagesNav with client navigation
Move sidebar page links into a client component for prefetch and
router.push while keeping the layout shell server-friendly.

Made-with: Cursor
